refactor #66: CartItem 컴포넌트 분리
