+++ fe/src/components/Main/Cart/Cart.module.css
@@ -42,35 +42,6 @@
   border-radius: 1rem;
 }
 
-.ItemContent {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 8px;
-}
-
-.ItemImage {
-  max-width: 100%;
-  height: auto;
-}
-
-.ItemCount {
-  position: absolute;
-  top: 0;
-  left: 0;
-  transform: translate(-25%, -25%);
-  width: 30%;
-  aspect-ratio: 1/1;
-  display: inline-flex;
-  justify-content: center;
-  align-items: center;
-  border-radius: 50%;
-  font-size: 1.5rem;
-  font-weight: bold;
-  color: white;
-  background-color: rgb(254, 193, 61);
-}
-
 .ModalContainer {
   position: absolute;
   left: 0;
